+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/MiniAppSdkConfig.kt
@@ -7,7 +7,7 @@ package com.rakuten.tech.mobile.miniapp
  * @property subscriptionKey Subscription Key for the Platform API.
  * @property hostAppVersionId Version of the host app, used to determine feature compatibility for Mini App.
  * @property hostAppUserAgentInfo User Agent information from Host App.
- * @property isTestMode Whether the sdk is making use of the Test API Endpoints for under "Testing" mini apps.
+ * @property isPreviewMode Whether the host app wants to use the API Endpoints under "Preview" mode.
  */
 data class MiniAppSdkConfig(
     val baseUrl: String,
@@ -15,11 +15,11 @@ data class MiniAppSdkConfig(
     val subscriptionKey: String,
     val hostAppVersionId: String,
     val hostAppUserAgentInfo: String,
-    val isTestMode: Boolean
+    val isPreviewMode: Boolean
 ) {
-    internal val key = "$baseUrl-$isTestMode-$rasProjectId-$subscriptionKey-$hostAppVersionId"
+    internal val key = "$baseUrl-$isPreviewMode-$rasProjectId-$subscriptionKey-$hostAppVersionId"
 
-    @Deprecated("Use rasProjectId instead of rasAppId")
+    @Deprecated("Use isPreviewMode instead of isTestMode, Use rasProjectId instead of rasAppId")
     constructor(
         baseUrl: String,
         rasAppId: String,
@@ -27,14 +27,15 @@ data class MiniAppSdkConfig(
         subscriptionKey: String,
         hostAppVersionId: String,
         hostAppUserAgentInfo: String,
-        isTestMode: Boolean
+        isTestMode: Boolean,
+        isPreviewMode: Boolean = isTestMode
     ) : this(
         baseUrl = baseUrl,
         rasProjectId = rasProjectId,
         subscriptionKey = subscriptionKey,
         hostAppVersionId = hostAppVersionId,
         hostAppUserAgentInfo = hostAppUserAgentInfo,
-        isTestMode = isTestMode
+        isPreviewMode = isPreviewMode
     )
 
     init {