+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/MiniAppSdkConfigSpec.kt
@@ -9,13 +9,13 @@ class MiniAppSdkConfigSpec {
     fun `the key pattern should match the defined scheme`() {
         val config = MiniAppSdkConfig(
             baseUrl = TEST_URL_HTTPS_2,
-            isTestMode = true,
+            isPreviewMode = true,
             rasProjectId = TEST_HA_ID_PROJECT,
             subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,
             hostAppVersionId = TEST_HA_ID_VERSION,
             hostAppUserAgentInfo = TEST_HA_NAME
         )
-        config.key shouldEqual "${config.baseUrl}-${config.isTestMode}" +
+        config.key shouldEqual "${config.baseUrl}-${config.isPreviewMode}" +
                 "-${config.rasProjectId}-${config.subscriptionKey}-${config.hostAppVersionId}"
     }
 
@@ -25,12 +25,13 @@ class MiniAppSdkConfigSpec {
             baseUrl = TEST_URL_HTTPS_2,
             isTestMode = true,
             rasProjectId = TEST_HA_ID_PROJECT,
+            isPreviewMode = true,
             rasAppId = TEST_HA_ID_APP,
             subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,
             hostAppVersionId = TEST_HA_ID_VERSION,
             hostAppUserAgentInfo = TEST_HA_NAME
         )
-        config.key shouldEqual "${config.baseUrl}-${config.isTestMode}" +
+        config.key shouldEqual "${config.baseUrl}-${config.isPreviewMode}" +
                 "-${config.rasProjectId}-${config.subscriptionKey}-${config.hostAppVersionId}"
     }
 
@@ -45,7 +46,7 @@ class MiniAppSdkConfigSpec {
             hostAppVersionId = TEST_HA_ID_VERSION,
             hostAppUserAgentInfo = TEST_HA_NAME
         )
-        config.key shouldEqual "${config.baseUrl}-${config.isTestMode}" +
+        config.key shouldEqual "${config.baseUrl}-${config.isPreviewMode}" +
                 "-${config.rasProjectId}-${config.subscriptionKey}-${config.hostAppVersionId}"
     }
 
@@ -53,6 +54,8 @@ class MiniAppSdkConfigSpec {
     fun `should throw exception when api url is not https`() {
         MiniAppSdkConfig(
             baseUrl = "http://www.example.com/1",
+            isPreviewMode = false,
+            rasAppId = TEST_HA_ID_APP,
             isTestMode = false,
             rasProjectId = TEST_HA_ID_PROJECT,
             subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,
@@ -65,6 +68,8 @@ class MiniAppSdkConfigSpec {
     fun `should throw exception when api url is blank`() {
         MiniAppSdkConfig(
             baseUrl = " ",
+            isPreviewMode = true,
+            rasAppId = TEST_HA_ID_APP,
             isTestMode = true,
             rasProjectId = TEST_HA_ID_PROJECT,
             subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,
@@ -77,7 +82,7 @@ class MiniAppSdkConfigSpec {
     fun `should throw exception when projectId is blank`() {
         MiniAppSdkConfig(
             baseUrl = TEST_URL_HTTPS_2,
-            isTestMode = true,
+            isPreviewMode = true,
             rasProjectId = " ",
             subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,
             hostAppVersionId = TEST_HA_ID_VERSION,
@@ -89,6 +94,7 @@ class MiniAppSdkConfigSpec {
     fun `should throw exception when both projectId and rasAppId are blank`() {
         MiniAppSdkConfig(
             baseUrl = TEST_URL_HTTPS_2,
+            isPreviewMode = true,
             isTestMode = true,
             rasProjectId = " ",
             rasAppId = " ",
@@ -102,6 +108,8 @@ class MiniAppSdkConfigSpec {
     fun `should throw exception when subscriptionKey is blank`() {
         MiniAppSdkConfig(
             baseUrl = TEST_URL_HTTPS_2,
+            isPreviewMode = true,
+            rasAppId = TEST_HA_ID_APP,
             isTestMode = true,
             rasProjectId = TEST_HA_ID_PROJECT,
             subscriptionKey = " ",
@@ -114,6 +122,8 @@ class MiniAppSdkConfigSpec {
     fun `should throw exception when hostAppVersionId is blank`() {
         MiniAppSdkConfig(
             baseUrl = TEST_URL_HTTPS_2,
+            isPreviewMode = true,
+            rasAppId = TEST_HA_ID_APP,
             isTestMode = true,
             rasProjectId = TEST_HA_ID_PROJECT,
             subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,