+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/RealMiniAppSpec.kt
@@ -123,6 +123,8 @@ class RealMiniAppSpec {
             baseUrl = TEST_URL_HTTPS_2,
             isTestMode = true,
             rasProjectId = TEST_HA_ID_PROJECT,
+            isPreviewMode = true,
+            rasAppId = TEST_HA_ID_APP,
             subscriptionKey = TEST_HA_SUBSCRIPTION_KEY,
             hostAppVersionId = TEST_HA_ID_VERSION,
             hostAppUserAgentInfo = TEST_HA_NAME