+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/api/ApiClientSpec.kt
@@ -160,7 +160,7 @@ open class ApiClientSpec {
         downloadApi: DownloadApi = mockDownloadApi
     ) = ApiClient(
         retrofit = retrofit,
-        isTestMode = false,
+        isPreviewMode = false,
         hostProjectId = hostProjectId,
         hostAppVersionId = hostAppVersionId,
         requestExecutor = requestExecutor,